+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 - Added mining particles to Glowing Mushrooms (#32)
 - Added an alias to ModBlockEntityTypes for `redstone_interface` -> `basic_redstone_interface` to prevent data loss, so 1.5.0 is no longer a breaking change
+- The Escape Rope no longer duplicates itself
 
 # 1.5.0